backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.10.4
asyncpg==0.30.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
pillow==10.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PySocks==1.7.1
pytz==2024.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.32.1
wsproto==1.2.0
